+++ packages/auth/src/providers/cognito/utils/signInStore.ts
@@ -2,8 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 import { CognitoAuthSignInDetails } from '../types';
-
-import { ChallengeName } from './clients/CognitoIdentityProvider/types';
+import { ChallengeName } from '../../../foundation/factories/serviceClients/cognitoIdentityProvider/types';
 
 // TODO: replace all of this implementation with state machines
 interface SignInState {