+++ scripts/tsc-compliance-test/CHANGELOG.md
@@ -3,6 +3,34 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.52](https://github.com/aws-amplify/amplify-js/compare/tsc-compliance-test@0.1.51...tsc-compliance-test@0.1.52) (2024-09-04)
+
+**Note:** Version bump only for package tsc-compliance-test
+
+## [0.1.51](https://github.com/aws-amplify/amplify-js/compare/tsc-compliance-test@0.1.50...tsc-compliance-test@0.1.51) (2024-09-03)
+
+**Note:** Version bump only for package tsc-compliance-test
+
+## [0.1.50](https://github.com/aws-amplify/amplify-js/compare/tsc-compliance-test@0.1.49...tsc-compliance-test@0.1.50) (2024-08-26)
+
+**Note:** Version bump only for package tsc-compliance-test
+
+## [0.1.49](https://github.com/aws-amplify/amplify-js/compare/tsc-compliance-test@0.1.48...tsc-compliance-test@0.1.49) (2024-08-21)
+
+**Note:** Version bump only for package tsc-compliance-test
+
+## [0.1.48](https://github.com/aws-amplify/amplify-js/compare/tsc-compliance-test@0.1.47...tsc-compliance-test@0.1.48) (2024-08-15)
+
+**Note:** Version bump only for package tsc-compliance-test
+
+## [0.1.47](https://github.com/aws-amplify/amplify-js/compare/tsc-compliance-test@0.1.46...tsc-compliance-test@0.1.47) (2024-08-07)
+
+**Note:** Version bump only for package tsc-compliance-test
+
+## [0.1.46](https://github.com/aws-amplify/amplify-js/compare/tsc-compliance-test@0.1.45...tsc-compliance-test@0.1.46) (2024-08-05)
+
+**Note:** Version bump only for package tsc-compliance-test
+
 ## [0.1.45](https://github.com/aws-amplify/amplify-js/compare/tsc-compliance-test@0.1.44...tsc-compliance-test@0.1.45) (2024-07-23)
 
 **Note:** Version bump only for package tsc-compliance-test